admin/python/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
